migrate more registration methods
