+++ src/background/messenger/strict/registration.ts
@@ -57,6 +57,13 @@ import {
   launchAuthIntegration,
 } from "@/background/partnerIntegrations";
 import { getAvailableVersion } from "@/background/installer";
+import {
+  collectPerformanceDiagnostics,
+  initTelemetry,
+  pong,
+  recordEvent,
+  sendDeploymentAlert,
+} from "@/background/telemetry";
 
 expectContext("background");
 
@@ -110,6 +117,12 @@ declare global {
     LAUNCH_AUTH_INTEGRATION: typeof launchAuthIntegration;
 
     GET_AVAILABLE_VERSION: typeof getAvailableVersion;
+
+    PING: typeof pong;
+    COLLECT_PERFORMANCE_DIAGNOSTICS: typeof collectPerformanceDiagnostics;
+    RECORD_EVENT: typeof recordEvent;
+    INIT_TELEMETRY: typeof initTelemetry;
+    SEND_DEPLOYMENT_ALERT: typeof sendDeploymentAlert;
   }
 }
 
@@ -163,5 +176,11 @@ export default function registerMessenger(): void {
     LAUNCH_AUTH_INTEGRATION: launchAuthIntegration,
 
     GET_AVAILABLE_VERSION: getAvailableVersion,
+
+    PING: pong,
+    COLLECT_PERFORMANCE_DIAGNOSTICS: collectPerformanceDiagnostics,
+    RECORD_EVENT: recordEvent,
+    INIT_TELEMETRY: initTelemetry,
+    SEND_DEPLOYMENT_ALERT: sendDeploymentAlert,
   });
 }
