+++ src/bricks/transformers/extensionDiagnostics.ts
@@ -18,7 +18,7 @@
 import { TransformerABC } from "@/types/bricks/transformerTypes";
 import { type Schema } from "@/types/schemaTypes";
 import { getDiagnostics as collectFrameDiagnostics } from "@/contentScript/performanceMonitoring";
-import { collectPerformanceDiagnostics as collectExtensionDiagnostics } from "@/background/messenger/api";
+import { collectPerformanceDiagnostics as collectExtensionDiagnostics } from "@/background/messenger/strict/api";
 import { propertiesToSchema } from "@/utils/schemaUtils";
 
 class ExtensionDiagnostics extends TransformerABC {
