+++ src/contentScript/performanceMonitoring.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { ping } from "@/background/messenger/api";
+import { ping } from "@/background/messenger/strict/api";
 import { isContextInvalidatedError } from "@/errors/contextInvalidated";
 import { isLoadedInIframe } from "@/utils/iframeUtils";
 import { getSettingsState } from "@/store/settings/settingsStorage";
